+++ source/lib/utils/ResourceRequestCollector.js
